pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.22",
]
